# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LMS API Documentation",
        "description": "API documentation for LMS Smart Notebook project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/notes/show/{note}": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Get a note by its ID",
                "operationId": "f9994b72edcc2d75ab858506bb2f221b",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "description": "Note ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Note fetched successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "student_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "course_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "unit_number": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "lesson_title": {
                                                    "type": "string",
                                                    "example": "Intro to JWT"
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "My Note"
                                                },
                                                "content": {
                                                    "type": "object",
                                                    "example": {
                                                        "type": "doc",
                                                        "content": []
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Note not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Note not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/notes/{note}": {
            "put": {
                "tags": [
                    "Notes"
                ],
                "summary": "Update a note by its ID",
                "operationId": "36ce6f9253b773e27788a1ecb6aa4645",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "description": "Note ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Note Title"
                                    },
                                    "content": {
                                        "example": "{}"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note updated successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Notes"
                ],
                "summary": "Delete a note by its ID",
                "operationId": "98f04136407fb09cbd86682e2bbca42d",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "description": "Note ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Note deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Note not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Note not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to delete note",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to delete note"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/notes/{note}/export-pdf": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Export note as PDF",
                "description": "Exports the specified note as a PDF file and returns the PDF file for download.",
                "operationId": "exportNotePdf",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "description": "ID of the note to export",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF file",
                        "content": {
                            "application/pdf": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Note not found"
                    },
                    "500": {
                        "description": "Failed to generate PDF"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}